src: Rename "zns" directory to "zoned".
"zoned" is a more appropriate name for both HDDs and SSDs that support the zone
interface.  ZNS is an NVMe standard for zoned SSDs only.  Zoned HDDs, on the
other hand, use a SATA standard called ZBC/ZAC.
